+++ crates/vertigo/src/driver_module/js_value/js_json_struct.rs
@@ -13,9 +13,10 @@ use crate::{
 };
 
 use super::{
+    js_json_context::JsJsonContext,
     memory_block_read::MemoryBlockRead,
     memory_block_write::MemoryBlockWrite,
-    serialize::{JsJsonContext, JsJsonDeserialize, JsJsonSerialize},
+    serialize::{JsJsonDeserialize, JsJsonSerialize},
 };
 
 const PARAM_TYPE: u32 = 1;