+++ crates/vertigo/src/driver_module/js_value/serialize.rs
@@ -1,62 +1,10 @@
 use std::collections::{BTreeMap, HashMap};
-use std::rc::Rc;
 
 use crate::driver_module::js_value::js_json_struct::JsJsonNumber;
 use crate::driver_module::js_value::MapItem;
 
 use super::js_json_struct::JsJson;
-
-#[derive(Debug)]
-struct JsJsonContextInner {
-    parent: Option<Rc<JsJsonContextInner>>,
-    current: String,
-}
-
-#[derive(Clone, Debug)]
-pub struct JsJsonContext {
-    inner: Rc<JsJsonContextInner>,
-}
-
-impl JsJsonContext {
-    pub fn new(current: impl Into<String>) -> JsJsonContext {
-        Self {
-            inner: Rc::new(JsJsonContextInner {
-                parent: None,
-                current: current.into(),
-            }),
-        }
-    }
-
-    pub fn add(&self, child: impl ToString) -> JsJsonContext {
-        Self {
-            inner: Rc::new(JsJsonContextInner {
-                parent: Some(self.inner.clone()),
-                current: child.to_string(),
-            }),
-        }
-    }
-
-    pub fn convert_to_string(&self) -> String {
-        let mut path = Vec::new();
-        let mut current = self.inner.clone();
-
-        loop {
-            path.push(current.current.clone());
-
-            let Some(parent) = current.parent.clone() else {
-                return path.into_iter().rev().collect::<Vec<_>>().join(" -> ");
-            };
-
-            current = parent;
-        }
-    }
-}
-
-impl std::fmt::Display for JsJsonContext {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(&self.convert_to_string())
-    }
-}
+use super::JsJsonContext;
 
 pub trait JsJsonSerialize {
     fn to_json(self) -> JsJson;